% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{bf_error}
\alias{bf_error}
\title{bf_error}
\usage{
bf_error(message)
}
\arguments{
\item{message}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, 2018
}
