% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_f.R
\name{buffer_f}
\alias{buffer_f}
\title{Buffer points in XY space}
\usage{
buffer_f(foo, buffer, reps)
}
\arguments{
\item{foo}{a data.frame to select from with columns x, y}

\item{buffer}{the minimum distance between output points}

\item{reps}{the number of repetitions for the points selection}
}
\description{
Returns the original data table with buffered points removed.

Runs numerous iterations, as the random point selection can result in more/fewer output points.

1. Randomly select a single point
2. Remove points within 50km of that point
3. Randomly select of the remaining points
}
\note{
Taken from https://davidrroberts.wordpress.com/2015/09/25/spatial-buffering-of-points-in-r-while-retaining-maximum-sample-size/
}
\author{
David R. Roberts (2015)

Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
