% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pts.R
\name{filter_pts}
\alias{filter_pts}
\alias{filter_pts_reset}
\alias{filter_pts_resample}
\title{Update filter values}
\usage{
filter_pts(
  indata,
  metric,
  value,
  inlayer = NULL,
  id_fieldname = "idfield",
  byfield = FALSE,
  reverse = FALSE,
  samplesize = 1e+05,
  sid = "sid3",
  par1 = NA
)

filter_pts_reset(indata, filters = NA)

filter_pts_resample(indata)
}
\arguments{
\item{indata}{TODO}

\item{metric}{one between \code{f_rangev}, \code{f_rangey}, \code{f_zscorey}, \code{f_rbiasy}, \code{f_rangeq}, \code{f_pos}, \code{f_editmap}, \code{f_selpts}}

\item{value}{value or values to apply (depending on metric)}

\item{inlayer}{optional: \code{sf} of fields (see also \code{id_fieldname})}

\item{id_fieldname}{optional: name of the \code{inlayer} field containing unique ID of fields (default: \code{"idfield"})}

\item{byfield}{if FALSE (default), consider data as a unique field; if TRUE, iterate each filter on field \code{id_fieldname}}

\item{reverse}{if TRUE, the filter is applied reversed (this makes sense only for \code{f_editmap} and \code{f_selpts}). default is FALSE.}

\item{samplesize}{maximum size of the sample of the original data to work with (default: 100000; if NA: all the points)}

\item{sid}{character: name of the ID field to be used (\code{"uid"}, \code{"sid1"}, \code{"sid2"}, \code{"sid3"} or \code{"sid4"})}

\item{par1}{additional parameter for filtering (now used only for area to filter / not to filter NA values)}

\item{filters}{TODO}
}
\description{
\code{filter_pts} updates filter values (altering the original dataframe).

\code{filter_pts_reset} resets all filter values to FALSE (altering the original dataframe).

\code{filter_pts_resample} resets sampling order (altering the original dataframe).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
