% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guinterp.R
\name{guinterp}
\alias{guinterp}
\title{Main GUI for app gestione dati}
\usage{
guinterp(language = "en")
}
\arguments{
\item{language}{"en" (default) or "it" currently implemented}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\author{
Luigi Ranghetti, phD (2019) <ranghetti.l@irea.cnr.it>
}
