% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guinterp_process.R
\name{guinterp_process}
\alias{guinterp_process}
\title{Process a group of fields}
\usage{
guinterp_process(inputpts, inlayer, grid_path = NA, filtered = TRUE,
  id_fieldname = "idfield", interp_dir = tempdir(),
  interp_method = "krige", interp_res = 5, out_crs = NA,
  samplesize = 10000, buffer_radius = 15, parallel = TRUE,
  vgm = NA, merge = TRUE, overwrite = TRUE, .shiny_session = NULL,
  .shiny_pbar_id = NULL)
}
\arguments{
\item{inputpts}{TODO}

\item{inlayer}{SpatialPolygonsDataFrame of fields (see also `id_fieldname`)}

\item{grid_path}{optional: TODO}

\item{filtered}{logical: is `inputpts` already filtered? If TRUE (default),
no additional filter is applied; if FALSE, a standard automatic filter is applied.}

\item{id_fieldname}{optional: name of the `inlayer` field containing unique ID of fields (default: "idfield")}

\item{interp_dir}{directory where rasters are stores (default: temporary directory)}

\item{interp_method}{interpolation method ("krige" or "idw")}

\item{interp_res}{output raster resolution (numeric, in metres)}

\item{out_crs}{optional: CRS of output raster (default: CRS of inputpts)}

\item{samplesize}{maximum size of the sample of the original data to work with (default: 10000; if NA: all the points)}

\item{buffer_radius}{numeric: buffer (default: 15) to be applied internally to field borders.}

\item{parallel}{(optional) Logical or integer: if TRUE (default), interpolation
is executed using multiple cores in order to speed up the execution.
The number of cores is automatically determined; specifying it is also
possible (e.g. `parallel = 4`).
If FALSE, the interpolation is forced to run with a single core.}

\item{vgm}{named list of variograms (names must correspond to values of field `id_fieldname`).
If NA (default), they are automatically computed.}

\item{merge}{logical: if TRUE (default), merge raster of single fields and
return the pèath of the merged raster; if FALSE, return the paths of the
single field rasters.}

\item{overwrite}{Logical: should existing raster being reprocessed? (default: FALSE)}

\item{.shiny_session}{TODO}

\item{.shiny_pbar_id}{TODO}
}
\description{
This function calls other functions of the library to perform
 interpolation over different fields.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
