% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{h}
\alias{h}
\alias{ht}
\alias{ph}
\title{Convert to HTML entities}
\usage{
h(t)

ht(x, i18n)

ph(...)
}
\arguments{
\item{x}{Character vector or list}

\item{i18n}{Object containing translations}

\item{...}{Elements to be pasted.}
}
\value{
`t()` and `ht()`: if `x` is a 1-length character, an `HTML` object;
 if `x` is a vector with length > 1, a list of `HTML` objects;
 if `x` is a list of vectors, a list of objects.

 `ph()`: an `HTML` element.
}
\description{
Internal functions used to convert  elements to `HTML`.

 `t()` is converts [list of] character vectors to `HTML`.

 `ht()` converts a vector of translation labels to a list of `HTML`
 translated entities.

 `ph()` pastes `HTML` entities.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2020) \email{ranghetti.l@irea.cnr.it}
}
