% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_par.R
\name{krige_par}
\alias{krige_par}
\title{Parallelise kriging}
\usage{
krige_par(
  formula,
  locations,
  newdata,
  model = NA,
  n_cores = NA,
  method = "krige",
  nmax = Inf,
  maxdist = Inf
)
}
\arguments{
\item{formula}{TODO}

\item{locations}{TODO}

\item{newdata}{TODO}

\item{model}{TODO}

\item{n_cores}{TODO}

\item{method}{TODO}

\item{nmax}{TODO}

\item{maxdist}{TODO}
}
\description{
TODO
}
\note{
Based on https://gis.stackexchange.com/questions/237672/how-to-achieve-parallel-kriging-in-r-to-speed-up-the-process

License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
