% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interp_grid.R
\name{make_interp_grid}
\alias{make_interp_grid}
\title{Create global output grid}
\usage{
make_interp_grid(indata_sf, outres = 5, outcrs = st_crs(indata_sf),
  offset = c(0, 0), border = 15)
}
\arguments{
\item{indata_sf}{TODO all input points}

\item{outres}{`numeric (1)` TODO spatial resolution of output raster (in crs unit)}

\item{outcrs}{`crs` output crs}

\item{border}{`numeric (1)` TODO extent of border (in crs units)}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
