% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_check.R
\name{path_check}
\alias{path_check}
\title{path_check}
\usage{
path_check(path)
}
\arguments{
\item{path}{\code{string} full path to a folder}
}
\description{
accessory functions to check that a directory exists and
is writeable
}
\author{
Luigi Ranghetti, phD (2019)
}
