% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_check.R
\name{raster_check}
\alias{raster_check}
\title{raster_check}
\usage{
raster_check(path)
}
\arguments{
\item{path}{\code{string} full path to a raster}
}
\description{
accessory functions to check that a raster exists and
has a correct grid
}
\author{
Luigi Ranghetti, phD (2019)
}
