% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputpts.R
\name{read_inputpts}
\alias{read_inputpts}
\title{Read yield data}
\usage{
read_inputpts(input, borders = NULL, varname = NA,
  .shiny_session = NULL, .shiny_pbar_id = NULL)
}
\arguments{
\item{input}{TODO}

\item{borders}{TODO}

\item{.shiny_session}{TODO}

\item{.shiny_pbar_id}{TODO}

\item{format}{`character (1)` TODO}

\item{year}{TODO}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
