% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_modal_message.R
\name{show_modal_message}
\alias{show_modal_message}
\title{Show modal message}
\usage{
show_modal_message(
  message,
  title = NULL,
  icon = "cog",
  icon_colour = "default",
  size = "s",
  easyClose = FALSE,
  footer = NULL
)
}
\arguments{
\item{message}{TODO}

\item{title}{TODO}

\item{icon}{TODO}

\item{icon_colour}{TODO}

\item{size}{TODO}

\item{easyClose}{TODO}

\item{footer}{TODO}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
