% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_buffer_m.R
\name{st_buffer_m}
\alias{st_buffer_m}
\title{Buffer longitude-latitude polygons in metres}
\usage{
st_buffer_m(x, dist)
}
\arguments{
\item{x}{TODO}

\item{dist}{TODO}
}
\description{
TODO
}
\note{
Modified from https://gis.stackexchange.com/questions/121489/1km-circles-around-lat-long-points-in-many-places-in-world

License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
