% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_crs_utm.R
\name{st_crs_utm}
\alias{st_crs_utm}
\alias{st_crs_utm_from_lonlat}
\alias{st_transform_utm}
\title{Functions to manage UTM projections}
\usage{
st_crs_utm(x)

st_crs_utm_from_lonlat(lon, lat = 45)

st_transform_utm(x)
}
\arguments{
\item{x}{TODO}

\item{lon}{TODO}

\item{lat}{TODO}
}
\description{
`st_crs_utm` gives the UTM zone of an `sf` object

`st_crs_utm_from_lonlat` gives the UTM zone of a couple of geographic coordinates

`st_transform_utm` transforms an `sf` object in the proper UTM zone
}
\note{
License: GPL 3.0

 See also \url{https://stackoverflow.com/questions/40132542/get-a-cartesian-projection-accurate-around-a-lat-lng-pair}.
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
