% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pts.R
\name{filter_pts}
\alias{filter_pts}
\alias{filter_pts_reset}
\alias{filter_pts_resample}
\title{Update filter values}
\usage{
filter_pts(
  indata,
  metric,
  value,
  inlayer = NULL,
  id_fieldname = "idfield",
  byfield = FALSE,
  reverse = FALSE,
  samplesize = 1e+05,
  par1 = NA
)

filter_pts_reset(indata, filters = NA)

filter_pts_resample(indata)
}
\arguments{
\item{indata}{TODO}

\item{metric}{one between f_rangev, f_rangey, f_zscorey, f_rbiasy, f_rangeq, f_pos, f_editmap}

\item{value}{value or values to apply (depending on metric)}

\item{inlayer}{optional: sf of fields (see also `id_fieldname`)}

\item{id_fieldname}{optional: name of the `inlayer` field containing unique ID of fields (default: "idfield")}

\item{byfield}{if FALSE (default), consider data as a unique field; if TRUE, iterate each filter on field `id_fieldname`}

\item{reverse}{if TRUE, the filter is applied reversed (this makes sense only for `f_editmap`). default is FALSE.}

\item{samplesize}{maximum size of the sample of the original data to work with (default: 100000; if NA: all the points)}

\item{par1}{additional parameter for filtering (now used only for area to filter / not to filter NA values)}

\item{filters}{TODO}
}
\description{
`filter_pts` updates filter values (altering the original dataframe).

`filter_pts_reset` resets all filter values to FALSE (altering the original dataframe).

`filter_pts_resample` resets sampling order (altering the original dataframe).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
