% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guinterp.R
\name{guinterp}
\alias{guinterp}
\title{Main GUI for app gestione dati}
\usage{
guinterp()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
