% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guinterp.R
\name{guinterp}
\alias{guinterp}
\title{Launch GUInterp}
\usage{
guinterp(language = "en")
}
\arguments{
\item{language}{Interface language (currently only English -- `"en"` --
and Italian -- `"it"` -- were implemented).
Default is to retrieve it from system settings.}
}
\value{
NULL (the function is called for its side effect)
}
\description{
Open the GUInterp graphical interface
}
\examples{
\dontrun{
guinterp()
}
}
\author{
Luigi Ranghetti, phD (2020) <ranghetti.l@irea.cnr.it>
}
