% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{h}
\alias{h}
\alias{ht}
\alias{ph}
\title{Convert to HTML entities}
\usage{
h(t)

ht(x, i18n)

ph(...)
}
\arguments{
\item{x}{Character vector or list}

\item{i18n}{Object containing translations}

\item{...}{Elements to be pasted.}
}
\value{
\code{t()} and \code{ht()}: if \code{x} is a 1-length character, an \code{HTML} object;
if \code{x} is a vector with length > 1, a list of \code{HTML} objects;
if \code{x} is a list of vectors, a list of objects.

\code{ph()}: an \code{HTML} element.
}
\description{
Internal functions used to convert  elements to \code{HTML}.

\code{t()} is converts \link{list of} character vectors to \code{HTML}.

\code{ht()} converts a vector of translation labels to a list of \code{HTML}
translated entities.

\code{ph()} pastes \code{HTML} entities.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2020) \email{ranghetti.l@irea.cnr.it}
}
