% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputpts_to_sf.R
\name{inputpts_to_sf}
\alias{inputpts_to_sf}
\title{Make `sf` from filtered data}
\usage{
inputpts_to_sf(indata, outcrs = 3857, all = FALSE, sid = "sid3")
}
\arguments{
\item{indata}{TODO}

\item{outcrs}{`CRS` output CRS}

\item{all}{logical (TRUE for all points, FALSE for only filtered ones)}

\item{sid}{character: name of the ID field to be used (`"uid"`, `"sid1"`, `"sid2"`, `"sid3"` or `"sid4"`)}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
