% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputpts_to_sf.R
\name{inputpts_to_sf}
\alias{inputpts_to_sf}
\title{Make \code{sf} from filtered data}
\usage{
inputpts_to_sf(indata, outcrs = 3857, all = FALSE, sid = "sid3")
}
\arguments{
\item{indata}{TODO}

\item{outcrs}{\code{CRS} output CRS}

\item{all}{logical (TRUE for all points, FALSE for only filtered ones)}

\item{sid}{character: name of the ID field to be used (\code{"uid"}, \code{"sid1"}, \code{"sid2"}, \code{"sid3"} or \code{"sid4"})}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
